% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{apa.desc}
\alias{apa.desc}
\title{give M and SD per group, with the 'apa' package, on the results of t_test()}
\usage{
apa.desc(t_test, x)
}
\arguments{
\item{t_test}{an apa::t_test object}

\item{x}{the name of a group in the apa::t_test object}
}
\description{
give M and SD per group, with the 'apa' package, on the results of t_test()
}
