% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{bayesvalues}
\alias{bayesvalues}
\title{print the standard BF criteria I use to the console}
\usage{
bayesvalues()
}
\value{
none. A print method is used instead to print the criteria to the console
}
\description{
print the standard BF criteria I use to the console
}
