% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{center_colmeans}
\alias{center_colmeans}
\title{center a variable}
\usage{
center_colmeans(x)
}
\arguments{
\item{x}{a matrix, or an object that can coerced to a matrix (vector)}
}
\value{
the centered object
}
\description{
faster than using scale() as per https://www.gastonsanchez.com/visually-enforced/how-to/2014/01/15/Center-data-in-R/
}
