% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{chi_prop_diff_report}
\alias{chi_prop_diff_report}
\title{report chi-square different of proportions}
\usage{
chi_prop_diff_report(var_d)
}
\arguments{
\item{var_d}{a chi-square different of proportions object}
}
\description{
report chi-square different of proportions
}
