% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{chomp}
\alias{chomp}
\title{Chomp off leading/trailing whitespace}
\usage{
chomp(x)
}
\arguments{
\item{x}{String to chomp()}
}
\description{
Much needed function to chomp the leading/trailing whitespace from a string.
}
\details{
just a wrapper around \code{gsub("^\\s+|\\s+$", "", <string>)}
}
\examples{
chomp(x = "  This is an unruly string  ")
}
