% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_render.R
\name{clean_render}
\alias{clean_render}
\title{Clean Rendering of Embedded RMarkdown Documents}
\usage{
clean_render(toRender = NULL, outDir = NULL, extension = ".pdf")
}
\arguments{
\item{toRender}{A string pertaining to the Rmd file to be rendered (no extension)}

\item{outDir}{The directory rendered reports should be saved to}

\item{extension}{A string pertaining to the extension to use for the knit report, default is .pdf.}
}
\description{
\code{clean_render} is a utility function for knitting embedded R Markdown documents. This
circumvents the write access error that typically occurs if you call `rmarkdown::render()`
from within an Rmarkdown document.
}

