% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cor.bf}
\alias{cor.bf}
\title{Calculate Pearson correlation, and also the corresponsind Bayes Factor, and sample size}
\usage{
cor.bf(data)
}
\arguments{
\item{data}{a data.frame, where only the first and the second columns will be used}
}
\value{
A list of the results of the statistical test
}
\description{
Calculate Pearson correlation, and also the corresponsind Bayes Factor, and sample size
}
