% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{cor_diff_report}
\alias{cor_diff_report}
\title{report cocor's different of correlations}
\usage{
cor_diff_report(cor_p)
}
\arguments{
\item{cor_p}{a cocor object}
}
\description{
report cocor's different of correlations
}
