% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{diffDate}
\alias{diffDate}
\title{Calculate the difference between dates (vectorized)}
\usage{
diffDate(chr.end.date, chr.start.date)
}
\arguments{
\item{chr.end.date}{End dates (in the same format), same length as chr.start.date}

\item{chr.start.date}{Start dates (in format YYYY-MM-DD). 
Strongly recommend using toDate to get this format}
}
\value{
Number of days (as numeric) between end date and start date
}
\description{
Calculate the difference between dates (vectorized)
}
\examples{
 
chr.start <- toDate(c('2017-01-01','2016-03-01','2015-01-01'))
chr.end <- toDate(c('2013-03-01', '2016-03-16', '2016-09-09'))
diffDate(chr.end, chr.start)
}
