% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{endwrite}
\alias{endwrite}
\title{Inserts some text at the end of the current line
adapted from https://githubmemory.com/repo/rstudio/rstudioapi/issues/218
Note that using ctrl+enter to run this would result in the text being added the the end of the line BELOW.
Use alt+enter instead.}
\usage{
endwrite(sign)
}
\arguments{
\item{sign}{text to be added (a string vector of length 1)}
}
\value{
none
}
\description{
Inserts some text at the end of the current line
adapted from https://githubmemory.com/repo/rstudio/rstudioapi/issues/218
Note that using ctrl+enter to run this would result in the text being added the the end of the line BELOW.
Use alt+enter instead.
}
