% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandData.R
\name{expandData}
\alias{expandData}
\title{Expand data over missing value
This is particularly useful for graphing (where you wnant NAs to indicate missing data)
This can also be useful when calculating rolling means, etc}
\usage{
expandData(df, mxt.key.cols)
}
\arguments{
\item{df}{is a data.frame of data}

\item{mxt.key.cols}{is a character vector that specifies the columns for which you want all combinations of
OR mxt.key.cols is a named list, where the values of the list are additional levels to include}
}
\description{
Expand data over missing value
This is particularly useful for graphing (where you wnant NAs to indicate missing data)
This can also be useful when calculating rolling means, etc
}
\examples{
df <- data.frame(a = 1:4, b = c('A','B','C','D'), c = runif(4), stringsAsFactors=FALSE)
expandData(df, c('a','b'))
expandData(df, list(a = c(1:10), b=LETTERS))
}
