% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{falseposrate}
\alias{falseposrate}
\title{False positive classification}
\usage{
falseposrate(threshold, truth, fitted)
}
\arguments{
\item{threshold}{a value between 0 and 1 for the value at which an outcome is classified}

\item{truth}{a vector of [0 | 1] values representing the true classification}

\item{fitted}{a vector in the range [0,1] values representing the predicted classification}
}
\description{
Returns the rate at which false positives are identified in a classification problem given a certain threshold. Especially for logistic/poisson regression whether a model correctly classifies a binary outcome depends on the threshold.
}
\details{
This function is primarily used to calculate values for the Receiver Operating Characteristics (ROC) curve function.
}
\examples{
m0 = glm(outcome ~ predictor1 + predictor2, family = binomial(), data = train)
testprobs = predict(m0, newdata = test, type = "response")
truth = train$outcome
falseposrate(0.5, truth, testprobs)
# Most useful for understanding the rate of true positives across a range of thresholds
sapply(0:10 / 10, function(x) {falseposrate(x, truth = truth, fitted = testprobs)})
}
