% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{fitvalues}
\alias{fitvalues}
\title{print the standart fit criteria I use to the console}
\usage{
fitvalues()
}
\value{
none. A print method is used instead to print the criteria to the console
}
\description{
print the standart fit criteria I use to the console
}
