% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flag_vars.R
\name{flag_vars}
\alias{flag_vars}
\title{Flag Variables for Mismatches}
\usage{
flag_vars(x, y)
}
\arguments{
\item{x}{A vector of any type}

\item{y}{A vector of any type}
}
\value{
boolean True or False indicating if all elements of x are equivalent to y.
#'
}
\description{
\code{flag_vars} is a utility function for comparing two vectors for equality. Messages indicating locations of
mismatches are printed to the console.
}
\examples{
\dontrun{
flag_vars(x = 1:10, y = 1:10)
set.seed(77)
dat <- data.frame(V1 = sample(1:2, 10, replace = TRUE), V2 = sample(1:2, 10, replace = TRUE))
dat
flag_vars(x = dat$V1, y = dat$V2)
}
}

