% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flipPath}
\alias{flipPath}
\title{flip back-slash to double slahes, and vice-versa}
\usage{
flipPath(text)
}
\arguments{
\item{text}{a char string}
}
\value{
a char string x, with flipped slashes
}
\description{
flip back-slash to double slahes, and vice-versa
}
