% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_names}
\alias{get_names}
\title{Inserts the names of an object below the current cursor location, in a new line.
Note that using ctrl+enter to run this would result in the text being added TWO lines below.
Use alt+enter instead.}
\usage{
get_names(x)
}
\arguments{
\item{x}{data.frame, or any other object with named elements}
}
\value{
none
}
\description{
Inserts the names of an object below the current cursor location, in a new line.
Note that using ctrl+enter to run this would result in the text being added TWO lines below.
Use alt+enter instead.
}
