% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getDateSequence}
\alias{getDateSequence}
\title{Get a sequence of dates, inclusive}
\usage{
getDateSequence(chr.start.date, chr.end.date, int.by = 1)
}
\arguments{
\item{chr.start.date}{The start date (only one)}

\item{chr.end.date}{The end date (only one)}

\item{int.by}{The interval by which dates should be given}
}
\value{
A character vector with the intervening days
}
\description{
Get a sequence of dates, inclusive
}
\examples{
getDateSequence('2017-04-01', '2018-01-01', 7)
getDateSequence('2016-02-01','2016-03-01', 1)
}
