% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getMonth}
\alias{getMonth}
\title{Get the first day of a month for a vector of dates (as character strings)
Much easier and faster than using substr+paste}
\usage{
getMonth(chr.date)
}
\arguments{
\item{chr.date}{A character vector of dates}
}
\value{
The months in a date (character) format. E.g., 2016-01-01
}
\description{
Get the first day of a month for a vector of dates (as character strings)
Much easier and faster than using substr+paste
}
\examples{
getMonth(c('2013-03-12','2013-09-02'))
}
