% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getMonthName}
\alias{getMonthName}
\title{Get the associated month (as a character) from a series of dates. 
Faster and easier than getting otherwise}
\usage{
getMonthName(chr.date)
}
\arguments{
\item{chr.date}{A character vector of dates}
}
\value{
The (character) months
}
\description{
Get the associated month (as a character) from a series of dates. 
Faster and easier than getting otherwise
}
\examples{
getMonthName(c('2013-03-01','2014-02-01'))
}
