% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getSeason}
\alias{getSeason}
\title{Get the associated seasons (as a character) from a series of dates. 
Uses the monthly season definition (Dec-Feb is Winter, Mar-May is spring, etc)
Faster and easier than calculating otherwise}
\usage{
getSeason(chr.date)
}
\arguments{
\item{chr.date}{A character vector of dates}
}
\value{
The (character) seasons
}
\description{
Get the associated seasons (as a character) from a series of dates. 
Uses the monthly season definition (Dec-Feb is Winter, Mar-May is spring, etc)
Faster and easier than calculating otherwise
}
\examples{
getSeason(c('2013-03-01','2014-02-01'))
}
