% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getTrailingDays}
\alias{getTrailingDays}
\title{Get a sequence of trailing days, inclusive}
\usage{
getTrailingDays(chr.end.date, int.length)
}
\arguments{
\item{chr.end.date}{The end date (only one)}

\item{int.length}{The number of days}
}
\value{
A character vector of int.length, with the trailing series of days
}
\description{
Get a sequence of trailing days, inclusive
}
\examples{
getTrailingDays('2017-04-01', 12)
}
