% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{getYear}
\alias{getYear}
\title{Get the associated years (as numeric integers) from a series of dates. 
Much faster than using \code{data.table::year} or \code{as.numeric(substr(...))}}
\usage{
getYear(chr.date)
}
\arguments{
\item{chr.date}{A character vector of dates}
}
\value{
The (numeric) year
}
\description{
Get the associated years (as numeric integers) from a series of dates. 
Much faster than using \code{data.table::year} or \code{as.numeric(substr(...))}
}
\examples{
getYear(c('2013-03-01','2014-02-01'))
}
