% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%out\%}
\alias{\%out\%}
\alias{\%not_in\%}
\title{A not_in operator}
\usage{
a \%out\% b
}
\arguments{
\item{a}{an object to search if it's in b}

\item{b}{an object where to look for object a}
}
\value{
bolian. TRUE if a is not %in% b
}
\description{
A not_in operator
}
