% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.defined}
\alias{is.defined}
\title{check if an object is defined}
\usage{
is.defined(sym)
}
\arguments{
\item{sym}{an object to check}
}
\value{
bollian. TRUE if sym is defined, otherwise FALSE.
}
\description{
check if an object is defined
}
