% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lm_with_cor}
\alias{lm_with_cor}
\title{internal function for GGally's ggdou, to add a trend line}
\usage{
lm_with_cor(data, mapping, ..., method = "pearson")
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{...}{...}

\item{method}{method}
}
\description{
internal function for GGally's ggdou, to add a trend line
}
