% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{lq}
\alias{lq}
\title{Calculate location quotients}
\usage{
lq(d)
}
\arguments{
\item{d}{a data frame of arbitrary size

The location quotient is a proportion of proportions. The typical use case is comparing local proportions to some standard (e.g. national) proportion as a quick way of understanding how exaggerated local differences are. 

If the local proportion of households with children is 0.15 and the national proportion is 0.10 then the LQ is \code{0.15 / 0.10 = 1.5}. Which is to say that the local proportion is half again larger than the national proportion.

Will use the row/proportions as the local proportions and the column totals as the reference.}
}
\description{
Converts a data frame of counts to a data frame of location quotients.
}
\examples{
lq_frame = lq(data)
}
