% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_n.R
\name{mean_n}
\alias{mean_n}
\title{SPSS Style Means}
\usage{
mean_n(df, n)
}
\arguments{
\item{df}{An R dataframe object}

\item{n}{Minimum number of items with valid responses}
}
\value{
A vector of type numeric.
}
\description{
\code{mean_n} returns the row-wise average of a series of vectors. The n
indicates the minimimum number of items a particpant should have responded
to in order to be given a mean.
}
\examples{
\dontrun{
set.seed(77)
x <- data.frame(matrix(rnorm(n = 50), ncol = 5))
x[1,1:4] <- NA
mean_n(x, n = 2)
}
}
\seealso{
\code{\link{scoring}}
}

