% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mmisc.R
\docType{package}
\name{Mmisc-package}
\alias{Mmisc}
\alias{Mmisc-package}
\title{Mmisc: Some Helper Functions for Michael's Code}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
some helper functions Michael Pinus (michael.pinus AT gmail.com) wrote and gathered over the years.
}
\keyword{internal}
