% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{multi.tests}
\alias{multi.tests}
\title{Multiple statistical tests over a group of variables}
\usage{
multi.tests(fun = apa::t_test, df, vars, group.var, ...)
}
\arguments{
\item{fun}{A statistical functions, default is apa::t_test}

\item{df}{A data.frame containing `vars` and `group.var`}

\item{vars}{A char vector with the names of the dependent variables to test}

\item{group.var}{A string with the name of the}

\item{...}{additional parameters to pass to `fun`}
}
\value{
A list of the results of the statistical test
}
\description{
Multiple statistical tests over a group of variables
}
