% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{numbers2words}
\alias{numbers2words}
\title{convert numbers to literal numbers}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{a number in numeric format}
}
\description{
based on https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r
}
