% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{plot.avp}
\alias{plot.avp}
\title{Quick plot of fitted values as a function of actual}
\usage{
\method{plot}{avp}(m)
}
\arguments{
\item{model}{A model object of type \code{glm}}
}
\description{
X-Y scatterplot of fitted ~ actual with a line of equality
}
\examples{
plot.avp(mod.fit)
}
