% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot_trend_lines}
\alias{plot_trend_lines}
\title{a helper for GGally's ggpairs}
\usage{
plot_trend_lines(data, mapping, ...)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{...}{...}
}
\description{
a helper for GGally's ggpairs
}
