% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLast.R
\name{printLast}
\alias{printLast}
\title{print the last line, again, for use with keyboard shortcut}
\usage{
printLast()
}
\value{
none. a print method is used instead
}
\description{
print the last line, again, for use with keyboard shortcut
}
