% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{pvalue.correct}
\alias{pvalue.correct}
\title{convert numeric p-values to strings for reporting, without leading zero}
\usage{
pvalue.correct(p)
}
\arguments{
\item{p}{a p-value in numeric value}
}
\description{
convert numeric p-values to strings for reporting, without leading zero
}
