% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{pvalue.report}
\alias{pvalue.report}
\title{convert string p-values to strings for reporting, without leading zero}
\usage{
pvalue.report(p)
}
\arguments{
\item{p}{a p-value in string value}
}
\description{
convert string p-values to strings for reporting, without leading zero
}
