% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{pvalue.stars}
\alias{pvalue.stars}
\title{convert numeric p-values to asterik stars}
\usage{
pvalue.stars(p)
}
\arguments{
\item{p}{a p-value in numeric format}
}
\description{
convert numeric p-values to asterik stars
}
