% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick.csv}
\alias{quick.csv}
\title{quickly export a data.frame to a CSV file}
\usage{
quick.csv(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
nothing. output is saved as a csv file named as per the data.frame, in the working directroy
}
\description{
quickly export a data.frame to a CSV file
}
