% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick.xls}
\alias{quick.xls}
\title{quickly export a data.frame to an Excel XLSX file}
\usage{
quick.xls(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
nothing. output is saved as a xlsx file named as per the data.frame, in the working directroy
}
\description{
quickly export a data.frame to an Excel XLSX file
}
