% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_cor.bf}
\alias{report_cor.bf}
\title{report correlation with BF, created by cor.bf()}
\usage{
report_cor.bf(corObject, BF01 = F)
}
\arguments{
\item{corObject}{a cor.bf() object}

\item{BF01}{should the BF be 10, or 01 based}
}
\description{
report correlation with BF, created by cor.bf()
}
