% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{rmse.glm}
\alias{rmse.glm}
\title{Extract RMSE and pseudo RSquared}
\usage{
rmse.glm(m)
}
\arguments{
\item{model}{A model object of type \code{glm}}
}
\description{
Calculates Mean squared error and r-squared from the fitted and actual values. Returns each (respectively) in a 1x2 matrix.
}
\examples{
rmse.glm(mod.fit)
}
