% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{roclines}
\alias{roclines}
\title{Calculate x-y pairs of true/false positives for ROC charts}
\usage{
roclines(truth, fitted, thresholds = 0:100/100)
}
\arguments{
\item{truth}{a vector of [0 | 1] values representing the true classification}

\item{fitted}{a vector in the range [0,1] values representing the predicted classification}
}
\description{
Returns a matrix of true/false positive rates which can be plotted as a ROC curve.
}
\examples{
m0 = glm(outcome ~ predictor1 + predictor2, family = binomial(), data = train)
testprobs = predict(m0, newdata = test, type = "response")
roc.vals.train = roclines(train$outcome, m0$fitted)
roc.vals.test = roclines(train$outcome, testprobs)
}
