% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{row_means}
\alias{row_means}
\title{rowMeans that JUST WORKS inside a dplyr::mutate() call}
\usage{
row_means(..., na.rm = TRUE)
}
\arguments{
\item{...}{a list of unquoted variable names}

\item{na.rm}{how to handle NAs}
}
\value{
A meaned vector
}
\description{
just name the variable to average, and you shall have their mean
}
