% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMisc.R
\name{runner}
\alias{runner}
\title{Calculate a running summary}
\usage{
runner(vec, w, FUN)
}
\arguments{
\item{vec}{Vector to summarize}

\item{w}{window (length) for moving summary}

\item{FUN}{summary function}
}
\description{
Running summary (mean, sum, etc) of a vector over a moving window.
}
\examples{
# weekly moving average
runner(daily, w = 7, FUN = mean)
}
