% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv}
\alias{rv}
\title{Reverse code a scale rating}
\usage{
rv(vector, minValue = NA, maxValue = NA, keepAttr = FALSE)
}
\arguments{
\item{vector}{a vector to reverse code}

\item{minValue}{set the minimum value of the scale, if missing, minimum is estimated from the empirical minimum value (what's in the vector)}

\item{maxValue}{set the maximum value of the scale, if missing, maximum is estimated from the empirical maximum value (what's in the vector)}

\item{keepAttr}{bollian, whether to keep or discard the scale's attributes (labels)}
}
\value{
A reverse coded vector
}
\description{
Reverse code a scale rating vector with either specified or empirical minimum and maximum values, and while keeping or discarding the scale attributes (labels)
}
