% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv4}
\alias{rv4}
\title{Reverse code a 1-4 scale rating}
\usage{
rv4(vector)
}
\arguments{
\item{vector}{a vector to reverse code}
}
\value{
A reverse coded vector
}
\description{
Reverse code a 1-4 scale rating
}
