% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv4zero}
\alias{rv4zero}
\title{Reverse code a 0-3 scale rating}
\usage{
rv4zero(vector)
}
\arguments{
\item{vector}{a vector to reverse code}
}
\value{
A reverse coded vector
}
\description{
Reverse code a 0-3 scale rating
}
