% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv5}
\alias{rv5}
\title{Reverse code a 1-5 scale rating}
\usage{
rv5(vector)
}
\arguments{
\item{vector}{a vector to reverse code}
}
\value{
A reverse coded vector
}
\description{
Reverse code a 1-5 scale rating
}
