% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv5zero}
\alias{rv5zero}
\title{Reverse code a 0-4 scale rating}
\usage{
rv5zero(vector)
}
\arguments{
\item{vector}{a vector to reverse code}
}
\value{
A reverse coded vector
}
\description{
Reverse code a 0-4 scale rating
}
