% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rv7}
\alias{rv7}
\title{Reverse code a 1-7 scale rating}
\usage{
rv7(vector)
}
\arguments{
\item{vector}{a vector to reverse code}
}
\value{
A reverse coded vector
}
\description{
Reverse code a 1-7 scale rating
}
