% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sround.R
\name{sround}
\alias{sround}
\title{Simple Round}
\usage{
sround(x, n = 0)
}
\arguments{
\item{x}{A numeric vector to be rounded}

\item{n}{Number of decimals to round to (default is to integer)}
}
\value{
A vector of type numeric.
}
\description{
\code{sround} returns the row-wise average of a series of vectors. The n argument
indicates the minimimum number of items a particpant should have responded
to in order to be given an average.
}
\examples{
\dontrun{
x <- seq(from = 1, to = 5, by = .25)
round(x)
sround(x)
round(x, 1)
sround(x, 1)
}
}
\seealso{
\code{\link{scoring}}
}

