% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{theme_apa}
\alias{theme_apa}
\title{ggplot2 APA-style}
\usage{
theme_apa(base_size = 14, base_family = "", box = FALSE)
}
\arguments{
\item{base_size}{base_size}

\item{base_family}{base_family}

\item{box}{box}
}
\description{
theme modified version of papaja's theme_apa()
}
