% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toEXTERNAL}
\alias{toEXTERNAL}
\title{Open data.frame (or any R object) in an external app.}
\usage{
toEXTERNAL(x, app_path)
}
\arguments{
\item{x}{a data.frame/tibble}

\item{app_path}{a string with the path for the app}
}
\value{
none
}
\description{
Allows analysis in exteranl Stat apps, such as JASP or jamovi
}
