% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toJAMOVI}
\alias{toJAMOVI}
\title{Open data.frame/tibble in an jamovi}
\usage{
toJAMOVI(x)
}
\arguments{
\item{x}{a data.frame/tibble}
}
\value{
none
}
\description{
Allows analysis in jamovi. Dynamically sets the path to jamovi (might fail if more than one version is installed)
}
