% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{var_diff_report}
\alias{var_diff_report}
\title{report var.test's different of variances}
\usage{
var_diff_report(var_d)
}
\arguments{
\item{var_d}{a var.test's different of variances object}
}
\description{
report var.test's different of variances
}
