% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{row_sums}
\alias{row_sums}
\title{rowSums that JUST WORKS inside a dplyr::mutate() call}
\usage{
row_sums(...)
}
\arguments{
\item{...}{a list of unquoted variable names}
}
\value{
A summed vector
}
\description{
just name the variable to sum, and you shall have their sum
}
